healthcheck: move code to own file
